addTests

+++ src/main/java/com/phgnight/Drink.java
@@ -1,7 +1,7 @@
 package com.phgnight;
 
 public interface Drink {
-    public double getPrice();
+    public int getPrice();
     public String getDescription();
 
 }
